requirements.txt
efficientnet-pytorch==0.7.1
numpy==1.21.1
pandas==1.3.0
Pillow==8.3.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tqdm==4.61.2
typing-extensions==3.10.0.0
